in-degree zero shenanigans
The tasks which are ready to start are those whose blocking tasks are
completed (marked in green). This makes it difficult to see the tasks
with in-degree zero, which are ready to be started but do not have green
above them. So let's add a dummy green node above them.